--- Sources/StreamChatUI/ChatMessageList/Reactions/ChatMessageReactionsView+ItemView.swift
@@ -1,11 +1,0 @@
-//
-// Copyright © 2021 Stream.io Inc. All rights reserved.
-//
-
-import StreamChat
-import UIKit
-
-extension ChatMessageReactionsView {
-    @available(*, deprecated, message: "Use ChatMessageReactionItemView instead")
-    public typealias ItemView = ChatMessageReactionItemView
-}
